move DDTeamCollectionUnitTest to a separate file
